+++ example/example.dart
@@ -15,61 +15,61 @@ main() async {
   print('[main] looked up isolate: $lookedUp');
 
   // send a message to the isolate using the IsolateRef.
-  receiverRef.send('foo bar');
+  receiverRef.send('foo bar', type: 'string');
 }
 
 // the entry point for the receiver isolate
 receiver() {
-  print('[$context] receiver started');
+  print('[$isolateContext] receiver started');
 
   // counter for received messages
   int msgCount = 0;
 
   // register a listener for messages sent to this isolate.
-  context.onMessage.listen((message) async {
-    print('[${context.path}] message received: ${message}');
+  isolateContext.onMessage.listen((message) async {
+    print('[${isolateContext.path}] message received: ${message}');
 
     // reply message to sender
     message.replyTo?.send('re: ${message.content}');
 
     if (message.sender != null) {
       // look up sender by path
-      IsolateRef lookedUpSender = await context.lookupIsolate(message.sender.path);
-      print('[$context] looked up sender: $lookedUpSender');
+      IsolateRef lookedUpSender = await isolateContext.lookupIsolate(message.sender.path);
+      print('[$isolateContext] looked up sender: $lookedUpSender');
     }
     else {
-      IsolateRef spawnedIsolate = await context.spawnIsolate(new Uri(path: "/spawned"), spawned);
+      IsolateRef spawnedIsolate = await isolateContext.spawnIsolate(new Uri(path: "/spawned"), spawned);
       spawnedIsolate.send("hello spawned!");
     }
 
     // if 3 messages received, shutdown the cluster node.
     msgCount++;
     if (msgCount == 3) {
-      context.shutdownNode();
+      isolateContext.shutdownNode();
     }
   });
 
   // register a shutdown request listener
-  context.shutdownRequestListener = () {
-    print('[$context] shutdown requested!');
-    context.shutdownIsolate();
+  isolateContext.shutdownRequestListener = () {
+    print('[$isolateContext] shutdown requested!');
+    isolateContext.shutdownIsolate();
   };
 }
 
 sender() {
-  print('[${context}] sender started');
+  print('[${isolateContext}] sender started');
 
   // listen to replies from receiver
-  context.onMessage.listen((msg) => print('[${context.path}] message received: $msg'));
+  isolateContext.onMessage.listen((msg) => print('[${isolateContext.path}] message received: $msg'));
 
   // register a listener for isolate up events. a reference to the newly spawned isolate is provided to the listener. the listener
   // has access to the properties of the spawned isolate and can send messages to it via the IsolateRef.
-  context.onIsolateUp
+  isolateContext.onIsolateUp
       .where((ref) => ref.property('type') == 'receiver')
-      .listen((ref) => ref.send(context.property('msg')));
+      .listen((ref) => ref.send(isolateContext.property('msg')));
 }
 
 spawned() {
-  print('[$context] spawned started');
-  context.onMessage.listen((msg) => print('[$context] message received: $msg'));
+  print('[$isolateContext] spawned started');
+  isolateContext.onMessage.listen((msg) => print('[$isolateContext] message received: $msg'));
 }